Consolidate hooks.
