+++ client/src/hooks/useLanduseData.ts
@@ -1,6 +1,9 @@
+import { useMemo } from "react";
+import { useQuery } from "@tanstack/react-query";
 import { queryKeys } from "../api/queryKeys";
+import { fetchLanduse } from "../api/landuseApi";
+import { useWatershed } from "../contexts/WatershedContext";
 import { useLayerQuery } from "./useLayerQuery";
-import { useLanduseDataOnly } from "./useLanduseDataOnly";
 import type { LanduseMap } from "../api/types";
 
 export interface UseLanduseDataResult {
@@ -10,8 +13,15 @@ export interface UseLanduseDataResult {
 }
 
 export function useLanduseData(runId: string | null): UseLanduseDataResult {
-  const { landuseData, isLoading, landuseLegendMap, enabled } =
-    useLanduseDataOnly(runId);
+  const { layerDesired, isEffective } = useWatershed();
+  const enabled = Boolean(layerDesired.landuse.enabled && runId);
+  const landuseEffective = isEffective("landuse");
+
+  const { data: landuseData = null, isLoading } = useQuery<LanduseMap>({
+    queryKey: queryKeys.landuse.undisturbed(runId ?? ""),
+    queryFn: ({ signal }) => fetchLanduse({ runId: runId! }, signal),
+    enabled,
+  });
 
   useLayerQuery("landuse", {
     enabled,
@@ -19,6 +29,23 @@ export function useLanduseData(runId: string | null): UseLanduseDataResult {
     hasData: landuseData != null && Object.keys(landuseData).length > 0,
     queryKey: queryKeys.landuse.undisturbed(runId ?? ""),
   });
+
+  const landuseLegendMap = useMemo(() => {
+    if (
+      !landuseEffective ||
+      !landuseData ||
+      Object.keys(landuseData).length === 0
+    ) {
+      return {} as Record<string, string>;
+    }
+    const legend: Record<string, string> = {};
+    for (const { color, desc } of Object.values(landuseData)) {
+      if (color && desc && !(color in legend)) {
+        legend[color] = desc;
+      }
+    }
+    return legend;
+  }, [landuseEffective, landuseData]);
 
   return {
     landuseData: landuseData ?? undefined,
