+++ client/src/hooks/useSubcatchmentData.ts
@@ -1,7 +1,8 @@
+import { useQuery } from "@tanstack/react-query";
 import { queryKeys } from "../api/queryKeys";
 import { fetchSubcatchments } from "../api/api";
 import { useWatershed } from "../contexts/WatershedContext";
-import { useLayerData } from "./useLayerData";
+import { useLayerQuery } from "./useLayerQuery";
 
 export interface UseSubcatchmentDataResult {
   subcatchments: GeoJSON.FeatureCollection | undefined;
@@ -14,16 +15,18 @@ export function useSubcatchmentData(
   const { layerDesired } = useWatershed();
   const enabled = Boolean(layerDesired.subcatchment.enabled && runId);
 
-  const { data, isLoading } = useLayerData(
-    "subcatchment",
-    queryKeys.subcatchments.byRun(runId ?? ""),
-    (signal) => fetchSubcatchments(runId!, signal),
+  const { data, isLoading, error } = useQuery({
+    queryKey: queryKeys.subcatchments.byRun(runId ?? ""),
+    queryFn: ({ signal }) => fetchSubcatchments(runId!, signal),
     enabled,
-    {
-      hasDataFn: (d) => (d?.features?.length ?? 0) > 0,
-      cancellationKey: queryKeys.subcatchments.all,
-    },
-  );
+  });
+
+  useLayerQuery("subcatchment", {
+    enabled,
+    isLoading,
+    hasData: !error && (data?.features?.length ?? 0) > 0,
+    queryKey: queryKeys.subcatchments.all,
+  });
 
   return { subcatchments: data ?? undefined, subLoading: isLoading };
 }
